Display customer ID at bottom of page
Did notice that this changes with every page refresh even though I have it
set to @customer just like what I am passing for editing the customer but when
going to edit it appears to pass a single number such as '1' to the url to get
to the edit page. Not sure why at this point though. Will proably also move this
inside of the admin area